Factor out service to a module
