+++ paasd/src/main.rs
@@ -1,51 +1,7 @@
-use futures::stream::Stream;
-use paas_types::process_service_server as server_types;
-use paas_types::{
-    ExecRequest, ExecResponse, LogsRequest, LogsResponse, StatusRequest, StatusResponse,
-    StopRequest, StopResponse,
-};
-use std::pin::Pin;
-use tonic::{
-    transport::{Certificate, Identity, Server, ServerTlsConfig},
-    Request, Response, Status,
-};
+use tonic::transport::{Certificate, Identity, Server, ServerTlsConfig};
 
-#[derive(Clone)]
-struct ProcessService;
-
-impl ProcessService {
-    fn new() -> Self {
-        Self
-    }
-}
-
-#[tonic::async_trait]
-impl server_types::ProcessService for ProcessService {
-    type GetLogsStream =
-        Pin<Box<dyn Stream<Item = Result<LogsResponse, tonic::Status>> + Send + Sync + 'static>>;
-
-    async fn exec(&self, _req: Request<ExecRequest>) -> Result<Response<ExecResponse>, Status> {
-        Err(Status::unimplemented(""))
-    }
-
-    async fn get_logs(
-        &self,
-        _req: Request<LogsRequest>,
-    ) -> Result<Response<Self::GetLogsStream>, Status> {
-        Err(Status::unimplemented(""))
-    }
-
-    async fn get_status(
-        &self,
-        _req: Request<StatusRequest>,
-    ) -> Result<Response<StatusResponse>, Status> {
-        Err(Status::unimplemented(""))
-    }
-
-    async fn stop(&self, _req: Request<StopRequest>) -> Result<Response<StopResponse>, Status> {
-        Err(Status::unimplemented(""))
-    }
-}
+mod service;
+use service::make_server;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -61,10 +17,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .identity(identity)
         .client_ca_root(client_ca_cert);
 
-    let service = ProcessService::new();
     Server::builder()
         .tls_config(tls)?
-        .add_service(server_types::ProcessServiceServer::new(service))
+        .add_service(make_server())
         .serve("127.0.0.1:8443".parse()?)
         .await?;
     Ok(())
